[06:00] Manually updating the UI
To demonstrate how people may be updating the UI manually, a click event
is added into the codebehind of the screen.
The click event updates the TextBox to be the contents of "Hello " + the
TextBlock.
However, this is not a "good thing" as now our logic is tightly coupled
to the properties of the control, and to those specific controls.