[26:30] Bound View to Commands in the ViewModel
2 new properties were added to the ViewModel
ICommand IncrementCommand
ICommand DoubleCommand
The constructor then initialises the commands with the delegates that
they should call when invoked (in this case the Increment() and Double()
methods within the ViewModel).

The View is then updated so that the buttons use bindings to connect to
the ICommands exposed on the ViewModel.