[41:00] Added Icecreams to View & ViewModel
Created a ObservableCollection<Icecream> Icecreams; in the ViewModel.
The VM constructor then populates the Icecreams collection.
The View is updated to have a ListView on it which Databinds to the
Icecreams collection.
The ListView is just doing a .ToSting() on each object at the moment.
The next step will create a DataTemplate to make each object look nicer.